+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
     name: 'pwix:accounts-tools',
-    version: '1.0.1-rc',
+    version: '2.0.0-rc',
     summary: 'Accounts management tools',
     git: 'https://github.com/trychlos/pwix-accounts-tools',
     documentation: 'README.md'
@@ -23,13 +23,12 @@ Package.onTest( function( api ){
 });
 
 function configure( api ){
-    api.versionsFrom( '2.9.0' );
-    api.use( 'deanius:promise@3.1.0' );
+    api.versionsFrom([ '2.9.0', '3.0-rc.0' ]);
     api.use( 'ecmascript' );
     api.use( 'mongo' );
-    api.use( 'pwix:options@2.1.0' );
+    api.use( 'pwix:options@2.1.1' );
     api.use( 'reactive-var' );
-    api.use( 'tmeasday:check-npm-versions@1.0.2', 'server' );
+    api.use( 'tmeasday:check-npm-versions@1.0.2 || 2.0.0-beta.0', 'server' );
 }
 
 // NPM dependencies are checked in /src/server/js/check_npms.js