+++ ChangeLog.md
@@ -2,11 +2,13 @@
 
 ## ChangeLog
 
-### 1.0.1-rc
+### 2.0.0-rc
 
     Release date: 
 
-    - 
+    - Get rid of 'deanius:promise' dependency
+    - Change APIs to always return a Promise (todo #3) thus bumping the major version number
+    - Meteor 3.0 ready
 
 ### 1.0.0